requirements.txt
pytesseract==0.3.6
numpy==1.19.2
Pillow==8.0.1
